scrape/requirements.txt
Django==2.2.9
django-rest-framework-mongoengine==3.4.0
djangorestframework==3.11.0
mongoengine==0.18.2